Move svg icons to assets
