+++ web/package.json
@@ -5,7 +5,7 @@
     "scripts": {
         "build": "cross-env NODE_ENV=production node ./esbuild.config.js",
         "dev": "cross-env NODE_ENV=development node ./esbuild.config.js",
-        "svelte-check": "svelte-check --workspace ./src --tsconfig ../tsconfig.json --threshold warning --fail-on-warnings",
+        "svelte-check": "svelte-check --workspace ./src --threshold warning --fail-on-warnings",
         "eslint": "eslint --ext ts -c .eslintrc.js src",
         "prettier": "prettier --write *.{js,json} .prettierrc .eslintrc.js src",
         "lint": "npm-run-all svelte-check eslint prettier"
