+++ web/src/assets/icons.ts
@@ -1,0 +1,6 @@
+/// <reference types="./image-import" />
+
+export { default as hrIcon } from "./hr.svg";
+export { default as bracesIcon } from "./braces.svg"
+export { default as paragraphIcon } from "./paragraph.svg"
+export { default as strikeThroughIcon } from "./strike-through.svg"
