+++ web/src/assets/image-import.d.ts
@@ -1,0 +1,4 @@
+declare module "*.svg" {
+    const icon: string;
+    export default icon;
+}
